refactor: 💡 list ledger api without account book to fix the bug
of unbalanced total credit&debit amount

✅ Closes: #3721 #3490
