+++ src/lib/utils/repo/line_item.repo.ts
@@ -46,3 +46,51 @@ export async function getOneLineItemWithoutInclude(lineItemId: number) {
 
   return lineItem;
 }
+
+export async function getAllLineItemsInPrisma(
+  companyId: number,
+  startDate: number,
+  endDate: number,
+  isDeleted?: boolean
+) {
+  const startDateInSecond = setTimestampToDayStart(startDate);
+  const endDateInSecond = setTimestampToDayEnd(endDate);
+  const deletedAt = isDeleted ? { not: null } : { equals: null };
+  const lineItems = await prisma.lineItem.findMany({
+    where: {
+      voucher: {
+        companyId,
+        date: {
+          gte: startDateInSecond,
+          lte: endDateInSecond,
+        },
+        deletedAt,
+      },
+      deletedAt,
+    },
+    include: {
+      voucher: {
+        select: {
+          id: true,
+          date: true,
+          type: true,
+          no: true,
+        },
+      },
+      account: {
+        select: {
+          code: true,
+          name: true,
+          parentId: true,
+        },
+      },
+    },
+    orderBy: {
+      voucher: {
+        date: 'asc',
+      },
+    },
+  });
+
+  return lineItems;
+}
